renaming functions, component, folder to match Converter instead of portfolio now that portfolio exists in the MainNavBar
